# old/api/docs/swagger.yaml
definitions:
...
  github_com_nicolailuther_butter_internal_dto.CancelOrder:
    properties:
      order_id:
        type: integer
      reason:
        type: string
    required:
    - order_id
    type: object
...
  github_com_nicolailuther_butter_internal_dto.CreateDispute:
    properties:
      description:
        minLength: 30
        type: string
      order_id:
        type: integer
      reason_category:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.DisputeReasonCategory'
    required:
    - description
    - order_id
    - reason_category
    type: object
...
  github_com_nicolailuther_butter_internal_dto.ExtendDeadline:
    properties:
      additional_days:
        maximum: 30
        minimum: 1
        type: integer
      order_id:
        type: integer
      reason:
        type: string
    required:
    - additional_days
    - order_id
    - reason
    type: object
...
  github_com_nicolailuther_butter_internal_dto.ResolveDispute:
    properties:
      refund_amount:
        type: number
      resolution:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.DisputeResolution'
      resolution_notes:
        type: string
    required:
    - resolution
    - resolution_notes
    type: object
...
  github_com_nicolailuther_butter_internal_models.MarketplaceDispute:
    properties:
      _opener:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.User'
      _order:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceOrder'
      _resolver:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.User'
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      opened_by:
        type: integer
      order_id:
        type: integer
      reason_category:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.DisputeReasonCategory'
      refund_amount:
        type: number
      resolution:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.DisputeResolution'
      resolution_notes:
        type: string
      resolved_at:
        type: string
      resolved_by:
        type: integer
      status:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.DisputeStatus'
      updated_at:
        type: string
    type: object
...
  github_com_nicolailuther_butter_internal_responses.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    name: Default
    url: https://default.dev
  description: Backend service for Butter
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Butter API
  version: "1.0"
paths:
...
  /api/v1/marketplace/disputes:
    get:
      description: Retrieve all disputes (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: List of disputes
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceDispute'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Disputes
      tags:
      - marketplace
  /api/v1/marketplace/disputes/{id}/resolve:
    patch:
      consumes:
      - application/json
      description: Resolve a dispute (admin only)
      parameters:
      - description: Dispute ID
        in: path
        name: id
        required: true
        type: integer
      - description: Resolution details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.ResolveDispute'
      produces:
      - application/json
      responses:
        "200":
          description: Dispute details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceDispute'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve Dispute
      tags:
      - marketplace
...
  /api/v1/marketplace/orders/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cancel details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.CancelOrder'
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel Order
      tags:
      - marketplace
...
  /api/v1/marketplace/orders/{id}/complete:
    post:
      description: Manually complete an order (buyer only)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete Order
      tags:
      - marketplace
...
  /api/v1/marketplace/orders/{id}/disputes:
    get:
      description: Retrieve dispute for a specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dispute details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceDispute'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Dispute by Order
      tags:
      - marketplace
    post:
      consumes:
      - application/json
      description: Open a dispute for an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Dispute details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.CreateDispute'
      produces:
      - application/json
      responses:
        "200":
          description: Dispute details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceDispute'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Open Dispute
      tags:
      - marketplace
  /api/v1/marketplace/orders/{id}/extend-deadline:
    post:
      consumes:
      - application/json
      description: Request deadline extension (seller only)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Extension details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.ExtendDeadline'
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Extend Deadline
      tags:
      - marketplace
...
  /api/v1/marketplace/orders/{id}/revisions:
    get:
      description: Retrieve all revisions for an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of revisions
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceRevisionRequest'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Revisions by Order
      tags:
      - marketplace
    post:
      consumes:
      - application/json
      description: Request changes to a deliverable
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Revision request details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.CreateRevisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Revision request details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceRevisionRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request Revision
      tags:
      - marketplace
  /api/v1/marketplace/orders/{id}/start:
    post:
      consumes:
      - application/json
      description: Start working on an order (seller only)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start Order
      tags:
      - marketplace
  /api/v1/marketplace/orders/{id}/timeline:
    get:
      description: Retrieve complete timeline of events for an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of timeline events
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceOrderTimeline'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Order Timeline
      tags:
      - marketplace
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PushApiKey:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"